FloatListJsonConverter4D , StatsNames Ordering and refactored maxDD period

+++ src/WebServer/SqCoreWeb/projects/ChartGenerator/src/app/ChrtGenWs.cs
@@ -139,7 +139,7 @@ public class ChrtGenWs
                     if (itemDate < minStartDate)
                         minStartDate = itemDate; // MinStart Date of the portfolio's
                     chartVal.Dates.Add(item.x);
-                    chartVal.Values.Add((int)item.y);
+                    chartVal.Values.Add((float)item.y);
                 }
 
                 // Step 4: Filling the Stats data
